eval/key_elements_fidelity/requirements.txt
# OCR
paddleocr
paddlepaddle

# Vision and Text Models
transformers
torch
Pillow
evaluate
bert-score
